quest:
    https://ru.wowhead.com/quest=29883
    https://ru.wowhead.com/quest=24958
	https://www.wowhead.com/quest=29727
    https://ru.wowhead.com/quest=31743
	https://ru.wowhead.com/quest=29901
	https://ru.wowhead.com/quest=29920
	https://ru.wowhead.com/quest=29929
	https://ru.wowhead.com/quest=30150
	https://ru.wowhead.com/quest=29898
	https://ru.wowhead.com/quest=29716
	https://ru.wowhead.com/quest=30157
	https://ru.wowhead.com/quest=31771
    https://ru.wowhead.com/quest=29726
	https://ru.wowhead.com/quest=30933
    https://ru.wowhead.com/quest=29752
	https://ru.wowhead.com/quest=30186
	https://ru.wowhead.com/quest=29894
	https://ru.wowhead.com/quest=29950
	https://ru.wowhead.com/quest=30053
	https://ru.wowhead.com/quest=29932
	https://ru.wowhead.com/quest=31185
	https://ru.wowhead.com/quest=31186
	https://ru.wowhead.com/quest=31319
	https://ru.wowhead.com/quest=30050
	https://ru.wowhead.com/quest=31439
	https://ru.wowhead.com/quest=30271
	https://ru.wowhead.com/quest=31262
	https://ru.wowhead.com/quest=30269
	https://ru.wowhead.com/quest=31080
	https://ru.wowhead.com/quest=30480
	https://ru.wowhead.com/quest=30491
	https://ru.wowhead.com/quest=30927
	https://ru.wowhead.com/quest=30828

spell:
    warrior:
        https://mop-twinhead.twinstar.cz/?spell=143268
    shaman:
        https://www.wowhead.com/wotlk/ru/spell=51514
    prist:
        https://ru.wowhead.com/spell=119032
    rogue:
        https://ru.wowhead.com/spell=108208
    Racial:
	    https://www.wowhead.com/spell=69041
	Generic:
	    https://ru.wowhead.com/spell=123709

npc:
    https://www.wowhead.com/npc=66281
	https://www.wowhead.com/npc=59037
	https://ru.wowhead.com/npc=7232
	https://ru.wowhead.com/npc=56034

mehanic:
    Бойчовская Гильдия (Beta 0.1)
	Травничество Пандарии
	Горное дело Пандарии
	Исправлен пакет SMSG_TIME_SYNC_REQ (исправлен баг когда игрок в главзах друго игрока лагал при движениях, плохо синхронизировался)
	Доступны 8 и 9 ранги бойцовской гильдии

item:
    https://ru.wowhead.com/item=6948

loot:
    исправлен лут от использования https://ru.wowhead.com/spell=84193

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -8807,7 +8807,7 @@ void ObjectMgr::LoadGossipMenu()
 
         GossipMenus gMenu;
 
-        gMenu.entry             = fields[0].GetUInt16();
+        gMenu.entry             = fields[0].GetUInt32();
         gMenu.text_id           = fields[1].GetUInt32();
 
         if (!GetGossipText(gMenu.text_id))
@@ -8850,8 +8850,8 @@ void ObjectMgr::LoadGossipMenuItems()
 
         GossipMenuItems gMenuItem;
 
-        gMenuItem.MenuId                = fields[0].GetUInt16();
-        gMenuItem.OptionIndex           = fields[1].GetUInt16();
+        gMenuItem.MenuId                = fields[0].GetUInt32();
+        gMenuItem.OptionIndex           = fields[1].GetUInt32();
         gMenuItem.OptionIcon            = fields[2].GetUInt32();
         gMenuItem.OptionText            = fields[3].GetString();
         gMenuItem.OptionType            = fields[4].GetUInt8();
